Add contract service and SQL seed data
- Implemented contract management functions in contracts_service.py, including listing, retrieving, creating, updating, and deleting contracts.
- Updated schema.sql to change quantity_mwh and price_per_mwh fields to DOUBLE PRECISION.
- Added seed.sql to populate the contracts table with initial data for various energy types and their attributes.

+++ backend/sql/schema.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS contracts (
   id SERIAL PRIMARY KEY,
   energy_type VARCHAR(50) NOT NULL,
-  quantity_mwh NUMERIC(12, 2) NOT NULL,
-  price_per_mwh NUMERIC(10, 2) NOT NULL,
+  quantity_mwh DOUBLE PRECISION NOT NULL,
+  price_per_mwh DOUBLE PRECISION NOT NULL,
   delivery_start DATE NOT NULL,
   delivery_end DATE NOT NULL,
   location VARCHAR(80) NOT NULL,
